fix python3 compatible

+++ src/bitmap.py
@@ -141,7 +141,7 @@ class BitMap(object):
         """
         Construct BitMap from hex string
         """
-        bitstring = format(int(hexstring, 16), "0" + str(len(hexstring)/4) + "b")
+        bitstring = format(int(hexstring, 16), "0" + str(int(len(hexstring)/4)) + "b")
         return cls.fromstring(bitstring)
 
     @classmethod
